requirements.txt
Flask==2.0.2
pbkdf2==1.3
Pillow==9.0.0
pyaes==1.6.1
gunicorn==20.1.0